Move core game types to a types folder Move some types to their own files. No functional changes.
