+++ client/src/constants.ts
@@ -3,6 +3,7 @@
 // Imports
 import charactersInit from './charactersInit';
 import colorsInit from './colorsInit';
+import { START_CARD_RANK } from './game/types/constants';
 import suitsInit from './suitsInit';
 import variantsInit from './variantsInit';
 
@@ -26,48 +27,10 @@ export const ARROW_COLOR = {
 };
 export const FADE_TIME = 350; // In milliseconds
 
-// These constants much match their server-side counterparts
-export enum ActionType {
-  Play = 0,
-  Discard = 1,
-  ColorClue = 2,
-  RankClue = 3,
-}
-export enum ClueType {
-  Color = 0,
-  Rank = 1,
-}
-export enum ReplayActionType {
-  Turn = 0,
-  Arrow = 1,
-  Sound = 2,
-  HypoStart = 3,
-  HypoEnd = 4,
-  HypoAction = 5,
-  HypoBack = 6,
-  HypoToggleRevealed = 7,
-}
-export enum ReplayArrowOrder {
-  Deck = -1,
-  Clues = -2,
-  Pace = -3,
-  Efficiency = -4,
-  MinEfficiency = -5,
-}
-export enum StackDirection { // Used in the "Up or Down" and "Reversed" variants
-  Undecided = 0,
-  Up = 1,
-  Down = 2,
-  Finished = 3,
-}
-export const STACK_BASE_RANK = 0;
-export const UNKNOWN_CARD_RANK = 6;
-export const START_CARD_RANK = 7;
-export const MAX_CLUE_NUM = 8;
 export const SHUTDOWN_TIMEOUT = 30; // In minutes
 export const SUIT_REVERSED_SUFFIX = ' Reversed';
 
-// Objects representating JSON files
+// Objects representing JSON files
 const COLORS = colorsInit();
 const SUITS = suitsInit(COLORS);
 const VARIANTS = variantsInit(COLORS, SUITS, START_CARD_RANK);
