+++ client/src/game/types/Character.ts
