+++ client/src/game/types/ClueType.ts
@@ -1,0 +1,5 @@
+enum ClueType {
+  Color = 0,
+  Rank = 1
+}
+export default ClueType;
