+++ client/src/game/types/MsgClue.ts
@@ -1,7 +1,7 @@
 // MsgClue represents how a clue looks on the server
 // On the client, the color is a rich object
 
-import { ClueType } from '../../constants';
+import ClueType from './ClueType';
 
 // On the server, the color is a simple integer mapping
 export default class MsgClue {
