+++ client/src/game/types/ReplayActionType.ts
@@ -1,0 +1,11 @@
+enum ReplayActionType {
+  Turn = 0,
+  Arrow = 1,
+  Sound = 2,
+  HypoStart = 3,
+  HypoEnd = 4,
+  HypoAction = 5,
+  HypoBack = 6,
+  HypoToggleRevealed = 7
+}
+export default ReplayActionType;
