+++ client/src/game/types/ReplayArrowOrder.ts
@@ -1,0 +1,9 @@
+enum ReplayArrowOrder {
+  Deck = -1,
+  Clues = -2,
+  Pace = -3,
+  Efficiency = -4,
+  MinEfficiency = -5
+}
+
+export default ReplayArrowOrder;
