+++ client/src/game/types/SimpleCard.ts
