+++ client/src/game/types/StackDirection.ts
@@ -1,0 +1,8 @@
+enum StackDirection {
+  Undecided = 0,
+  Up = 1,
+  Down = 2,
+  Finished = 3
+}
+
+export default StackDirection;
