+++ client/src/game/types/Variant.ts
