--- client/src/game/ui/ClockData.ts
@@ -1,5 +1,0 @@
-export default interface ClockData {
-  times: number[];
-  active: number;
-  timeTaken: number;
-}
