+++ client/src/game/ui/LearnedCard.ts
@@ -1,5 +1,5 @@
 // Imports
-import Suit from '../../Suit';
+import Suit from '../types/Suit';
 
 export default interface LearnedCard {
   suit: Suit | null;
