+++ client/src/game/ui/PlayStack.ts
@@ -3,7 +3,7 @@
 
 // Imports
 import Konva from 'konva';
-import { STACK_BASE_RANK } from '../../constants';
+import { STACK_BASE_RANK } from '../types/constants';
 import globals from './globals';
 import HanabiCard from './HanabiCard';
 import LayoutChild from './LayoutChild';
